Add files via upload
Diagram for Model dessign in a 'locale i18n standard', 'load-on-demmand', application. The models are separated to bring freedom and speed in choice of different languages and media associated for each instance. Features: easy load and queries, and make possible a step-by-step building of web objects, according to its bit weight.